Fix 32-bit compatibility: DELTA_MIN_FILE_SIZE u64, install-daemon.sh armv7
Changes DELTA_MIN_FILE_SIZE from usize to u64 so file size comparisons
don't silently truncate on 32-bit systems. Also adds armv7l/armv7 arch
mapping to install-daemon.sh.

+++ crates/hop-core/src/proto/mod.rs
@@ -292,7 +292,7 @@ pub const DEFAULT_PARALLEL_STREAMS: usize = 4;
 pub const DELTA_BLOCK_SIZE: usize = TRANSFER_CHUNK_SIZE;
 
 /// Minimum file size to consider for delta transfer.
-pub const DELTA_MIN_FILE_SIZE: usize = 64 * 1024;
+pub const DELTA_MIN_FILE_SIZE: u64 = 64 * 1024;
 
 /// Request to initiate a file transfer session.
 #[derive(Debug, Clone, Serialize, Deserialize)]
